+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0]
+### Added
+- Support for estimated environmental impact api request
+
+### Fixed
+- Outdated example in Readme.md
+
 ## [0.1.0]
 ## [0.1.0-beta.3]
 ### Added